move configuration.go
